requirements.txt
streamlit
PyPDF2
python-docx
langchain
langchain-community
chromadb
requests
numpy
scipy
scikit-learn
openai
google-generativeai
